Move AddEditStorageModal into storages page components directory
