+++ src/app/home/pages/StoragesPage/components/AddEditStorageModal/ProviderForms/AzureForm.tsx
@@ -17,10 +17,10 @@ import {
   addEditButtonText,
   isAddEditButtonDisabled,
   isCheckConnectionButtonDisabled,
-} from '../../../../common/add_edit_state';
-import ConnectionStatusLabel from '../../../../common/components/ConnectionStatusLabel';
+} from '../../../../../../common/add_edit_state';
+import ConnectionStatusLabel from '../../../../../../common/components/ConnectionStatusLabel';
 import { withFormik, FormikProps } from 'formik';
-import utils from '../../../../common/duck/utils';
+import utils from '../../../../../../common/duck/utils';
 
 /* 
 This URL is swapped out with downstream build scripts to point to the downstream documentation reference
